refactor: переместить компонент Footer в виджеты и обновить стили
