--- frontend/src/components/footer/footer.tsx
@@ -1,9 +1,0 @@
-import "./footer.css";
-
-export default function Footer() {
-	return (
-		<>
-			<footer className={"footer-footer"}></footer>
-		</>
-	);
-}
