+++ frontend/src/shared/assets/FooterImg.png
