+++ frontend/src/widgets/footer/ui/footer.module.css
@@ -1,5 +1,5 @@
-.footer-footer {
-	background-image: url("../../assets/FooterImg.png");
+.footer {
+	background-image: url("../../../shared/assets/FooterImg.png");
 	padding: 0;
 	margin: 0;
 	width: 100vw;
