+++ frontend/src/widgets/footer/ui/footer.tsx
@@ -1,0 +1,7 @@
+import styles from "./footer.module.css";
+
+export default function Footer() {
+	return (
+		<footer className={styles.footer}></footer>
+	);
+}
